Add support for Sly, step 3.
Refactor log4slime.lisp and its clone log4elisp.lisp, which was copied
in step 0.

log4elisp.lisp is left with the bulk of the code originally in
log4slime.lisp.  It defines the new package #:log4cl.elisp.

Where log4slime.lisp called Swank backend directly, log4elisp.lisp
defines generic functions that dispatch to methods in log4slime.lisp,
which makes the function calls to Swank.

The two elisp interfaces: ‘emacs-helper’ and ’get-buffer-log-menu’
are split between log4elisp.lisp and log4slime.lisp.  The bulk of the
code is in the former.  While the remnants in log4slime.lisp are
wrapper functions, providing log4elisp.lisp with a dynamic dispatch
parameter to reach back to methods associated with Swank.

log4slime.lisp also retains the top level forms that manipulate Swank.
Specifically, those are ‘swank::compile-string-for-emacs’ and
‘#:register-thread-stopper’.

Common functions that lookup the values for ‘*quicklisp-home*’ and
‘*system-directory*’ are left in log4elisp.lisp.  The symbols are
imported and re-exported by log4slime.lisp.

Finally the ‘log4slime:install’ function has been refactored so that
it can be shared with what will be log4sly.lisp.  Most of what was
‘install’ is now part of log4elisp.lisp.  What remains in
log4slime.lisp is a wrapper that passes the package name as a string
to ‘log4cl.elisp:install’.