Add support for SLY, step 2.
This replaces 12 references in log4cl.el that directly call SLIME
backend functions with generic functions.  The generic functions
dispatch on the variable ‘log4slime-backend’ which is set to
‘:log4slime’ by ‘log4slime-mode’.

For each generic function in log4cl.el, define a method in
log4slime.el which calls the appropriate function in slime.el.

Note: the log4slime: package name has been removed from the calls to
‘log4slime-eval’.  The appropriate package name is reapplied in the
‘log4slime-eval’ method.

This enables subsequent steps, which implements a log4sly.el package
that provides:

- log4sly-mode
- log message highlighting for the SLY repl.
- the corresponding 12 methods to call into the SLY backend.